Ensures all teams are seeded the same way

+++ .ipynb_checkpoints/supervisor-checkpoint.py
@@ -1,6 +1,4 @@
 import time
-from datetime import datetime
-
 from tasks import start_workers
 from tpg.model import Model
 from parameters import Parameters
@@ -13,61 +11,96 @@ from tpg.mutator import Mutator
 
 if __name__ == '__main__':
 
+    print("Connecting to the database...")
+
     Database.connect(
         user="postgres",
         password="template!PWD",
-        host="192.168.4.196",
+        host=Parameters.DATABASE_IP,
         port=5432,
         database="postgres"
     )
 
+    print("Database connected.")
+
     Database.clear()
     Database.load()
-    Database.connect_duckdb()
+    Database.connect_duckdb(Parameters.DATABASE_IP)
+
+    print("Duckdb connected.")
 
     model = Model()
 
     # Update the database
-    for program in model.programPopulation:
-        Database.add_program(program)
     for team in model.teamPopulation:
         Database.add_team(team)
 
+        for program in team.programs:
+            Database.add_program(program, team)
 
     time_elapsed = 0.0
     benchmarking_data = []
-    for generation in range(1, 10+1):
-        start_time = datetime.now()
-
+    for generation in range(1, 200+1):
         print(f"Starting generation {generation}...")
 
         teams_per_worker = {
             "desktop": [str(id) for id in Database.get_root_teams()]
+      #      "raspberrypi": [str(id) for id in Database.get_root_teams()[240:300]],
+     #       "beelink": [str(id) for id in Database.get_root_teams()[300:360]],
+        }
+
+        worker_batch_sizes = {
+            "desktop": 12,
+      #      "raspberrypi": 4,
+      #      "beelink": 4,
+      #      "macbook": 8
         }
 
         # Collect training information
         start_workers(
             teams_per_worker,
+            worker_batch_sizes,
             generation,
             model
         )
 
         for team in model.teamPopulation:
+            # We only purge root teams, otherwise, there would never be any surviving child teams.
+            if team.id not in [str(team_id) for team_id in Database.get_root_teams()]:
+                print("Leaf team detected. Not going to delete this one.")
+                continue
             if team.id not in model.get_survivor_ids(generation):
+                if team.luckyBreaks > 0:
+                    team.luckyBreaks -= 1
+                    Database.update_team(team, team.luckyBreaks)
+                    continue
+
                 for program in team.programs:
-                    Database.remove_program(program)
+                    Database.remove_program(program, team)
                 Database.remove_team(team)
+
+                for _team in model.teamPopulation:
+                    if str(team.id) == _team.id:
+                        model.teamPopulation.remove(team)
+                        print(f"Removing team {team.id}")
 
         model.repopulate()
 
-                print(Database.get_ranked_teams(generation).sort_values('rank').head(10))
-
-        end_time = datetime.now()
-        delta_time = end_time - start_time
-        time_elapsed += delta_time.total_seconds()
-        benchmarking_data.append([generation,time_elapsed])
-
-    df = pd.DataFrame(benchmarking_data, columns=['generation', 'time_elapsed'])
-    df.to_csv("car_racing_12_envs.csv")
+        # Mark the lucky breaks now
+        n = 3
+        top_n_teams = Database.get_ranked_teams(generation).sort_values('rank').head(3)['team_id'].to_list()
+        for team in model.teamPopulation:
+            if str(team.id) in [ str(team_id) for team_id in top_n_teams ]:
+                team.luckyBreaks += 1
+                Database.update_team(team, team.luckyBreaks)
 
 
+        print(Database.get_ranked_teams(generation).sort_values('rank').head(10))
+
+        benchmarking_data = pd.DataFrame([{
+            "generation": generation,
+            "time": time.time()
+        }])
+
+        Database.store("time_monitor", benchmarking_data)
+        print(f"Finished generation {generation}...")
